Move some format files to format/ dir
