+++ src/components/Create/components/pages/ProjectToken/components/CustomTokenSettings/CustomTokenSettings.tsx
@@ -11,7 +11,7 @@ import { useAppSelector } from 'hooks/AppSelector'
 import useMobile from 'hooks/Mobile'
 import { useEditingDistributionLimit } from 'redux/hooks/EditingDistributionLimit'
 import { inputMustExistRule } from 'utils/antd-rules'
-import { formatAmount } from 'utils/formatAmount'
+import { formatAmount } from 'utils/format/formatAmount'
 import { MAX_DISTRIBUTION_LIMIT, MAX_MINT_RATE } from 'utils/v2v3/math'
 import * as ProjectTokenForm from '../../hooks/ProjectTokenForm'
 import { ProjectTokensFormProps } from '../../hooks/ProjectTokenForm'
